// src/types/api-key.ts
export interface ApiKeyPermissions {
  uploadImages: boolean;
  uploadText: boolean;
  uploadFiles: boolean;
  extractColors?: boolean;
}

export interface ApiKey {
  id: string;
  name: string;
  key: string;
  createdAt: string;
  expiresAt: string | null;
  createdBy: string;
  permissions: ApiKeyPermissions;
  lastUsed?: string;
}